feat(FN-3689): move get fee record controller and rename to match other controllers
