+++ portal-api/src/v1/controllers/utilisation-report-service/fee-record-corrections/fee-record-correction.controller.test.ts
@@ -1,9 +1,9 @@
 import httpMocks, { MockResponse } from 'node-mocks-http';
 import { AxiosResponse, HttpStatusCode, AxiosError } from 'axios';
 import { Response } from 'express';
-import { getFeeRecordCorrection, GetFeeRecordCorrectionRequest } from '.';
-import api from '../../../../api';
-import { aGetFeeRecordCorrectionResponseBody } from '../../../../../../test-helpers/test-data/get-fee-record-correction-response-body';
+import { getFeeRecordCorrection, GetFeeRecordCorrectionRequest } from './fee-record-correction.controller';
+import api from '../../../api';
+import { aGetFeeRecordCorrectionResponseBody } from '../../../../../test-helpers/test-data/get-fee-record-correction-response-body';
 
 jest.mock('../../../../api');
