+++ portal-api/src/v1/controllers/utilisation-report-service/fee-record-corrections/index.ts
@@ -1,2 +1,3 @@
 export * from './pending-corrections.controller';
 export * from './fee-record-correction-transient-form-data.controller';
+export * from './fee-record-correction.controller';
