+++ portal-api/src/v1/controllers/utilisation-report-service/index.ts
@@ -5,5 +5,4 @@ export { getDueReportPeriodsByBankId } from './due-report-periods.controller';
 export { getLastUploadedReportByBankId } from './last-uploaded.controller';
 export { getNextReportPeriodByBankId } from './next-report-period.controller';
 export { validateUtilisationReportData } from './validate-utilisation-report-data.controller';
-export { getFeeRecordCorrection } from './fee-record-corrections/get-fee-record-correction.controller';
 export * from './fee-record-corrections';
